Fix css class conflicts
On mounted apps that uses the dfe autocomplete the govuk all replaces
any customization of the govuk base which is not the desired output.

Importing only what's needed can decrease the level of friction
with any apps that uses the library.

Also wrapped the css under the suggestions class so it also
doesn't cause conflicts with anything on other app. This needs
to be changed in the future.