Remove default_classes and use new attr setup
In govuk-components 3.0.5 the mechanism used to set default classes and
attributes changed so a single hash of attributes that includes a :class
key.

This change implements that here, default_classes are automatically
merged by the components gem with the supplied :html_attributes and made
available to the component via html_attributes.