Add note about inflections to the README
Find a lost TRN couldn't get the tests to pass in CI without adding an
acronym for the `DfE` namespace. It could be related to a quirk with how
the files in this gem are structured; in `app` as opposed to `lib`.

In the meantime, we should document this caveat.