Remove existing visual cue for suggestions
Currently, when displaying suggestions, this component uses a ::before
pseudo element containing the word 'Suggestions' to introduce the list.

In an audit, the GOV.UK Design System team found that having space
between the input and the suggestions meant that the input could be
confusing for users of screen magnifiers, since when zoomed into the
input the suggestions would not be visible and the user would not know
about the options.

This commit removes the existing 'Suggestions' UI and the gap between
the input and the suggestions. A subsequent commit will add a new visual
cue for the suggestions.