Add new visual cue to the input for suggestions
When the input has suggestions and the user has it focused, the input's
bottom border will appear 'shortened' so that it doesn't fill the whole
bottom of the input element. Combined with the removal of the gap
between the input and the suggestions in the previous commit, this makes
it visually apparent to the user that there is more content below the
input for them to consider, even if they are zoomed all the way into the
input.