Add sass field to package.json
The README.md for this repo recommends importing the sass directly from
node_modules. Some frontend build systems support importing Sass from a
path specified in the package.json.

This commit adds a sass field to the package.json to allow this, and
adds a brief note to the README.md so that users know they can do this.