Use @use to import govuk-frontend
The Sass @import syntax is deprecated. The replacement, @use, expects
dependencies to be explicitly imported, but at present this package's
styles just assume that govuk-frontend and its mixins will be available
on the page.

This commit uses @use to import the govuk styles and utilities.

To keep the changes simple, I've opted to import the whole
govuk-frontend package without a namespace.